requirements.txt
yfinance==0.2.43
ccxt==4.3.92
pandas==2.2.2
numpy==1.26.4
PyYAML==6.0.2
APScheduler==3.10.4
python-telegram-bot==21.4
ta==0.11.0
